Fix support for local echo suppression
When reading back what we have sent, we must expect that read returns
less than the full request in one read. The initial implementation
does not handle this correctly, thus it is buggy for larger requests
(e.g. ones which does not fit into the RX fifo of the hardware).

Let's fix this issue, but giving read reasonable parameters.

Signed-off-by: Michael Heimpold <michael.heimpold@in-tech.com>